requirements.txt
flask==2.2.2
flask-MySQLdb==1.0.1
flask-SocketIO==5.3.2
requests==2.28.1
gunicorn==20.1.0
mysql==0.0.3